DAE.eig() and Eigvalues mysc function updates
Update functions processing eigenvalues required during mode decomposition.

The functions allow to get a vector of (unsorted) eigenvalues, and cluster them into real, and complex conjugate pairs, returing additionally the original indices of the eigenvalues, to relate them with the appropriate eigenvectors.

For DAE and LatentModels we  moved the eig method to the LatentBaseModel class, as for any latent model with an evolution operator, the eig function is generic.