DAE.modes() implementation
This commit provides the functionality to perform the mode decomposition using a  `LatentBaseModel` instance.

The mode decomposition is applied to the evolution operator in latent space, and then a *linear* reprojection to the state space, ensures the decomposition of latent modes, can be transfered to the decomposition of state modes.