DAE mode decomposition operation. Brittle.
In this state of the code the dynamic mode decomposition using the DAE architecture is operational.

The mode decomposition hinges on the use of the linear decoder to map states from the latent space Z to the original state space X. The prediction performance of the DAE model using the linear decoder hinges strongly on the learned latent space. DAE architectures are very sensible to the initialization of the evolution operator, leading to poor prediction performance in the latent space (which in practice is mitigated by the non-linear decoder).

To get good performance, the user needs to initialize the evolution operator such that the frequencies of oscilation of the eigenvalues of the operator span a reasonable space of frequencies present in the dynamics of X.